Code cleanup:
Since we already have the define ISC_STATUS_LENGTH for the the length of
the status vector, we should use it! Replaced 'status[20]' with
'status[ISC_STATUS_LENGTH]'.

Replaced the different symbolic names for the maximum length of a path
(MAX_PATH, MAX_PATH_LENGTH, MAXPATHLEN) with one uniq symbol MAXPATHLEN.


git-svn-id: svn://svn.code.sf.net/p/firebird/code/firebird/trunk@1571 65644016-39b1-43b1-bf79-96bc8fe82c15